chore: Move directory for radio, checkbox
